Provide `git_repo` as a Pytest plugin
This way the `darker` and `graylint` Python packages can re-use it from
here instead of duplicating the implementation.
