+++ src/darkgraylib/tests/test_stdin_filename.py
@@ -11,7 +11,7 @@ import toml
 from darkgraylib import command_line
 from darkgraylib.config import BaseConfig, ConfigurationError
 from darkgraylib.git import RevisionRange
-from darkgraylib.tests.conftest import GitRepoFixture
+from darkgraylib.tests.plugin import GitRepoFixture
 from darkgraylib.tests.helpers import raises_if_exception
 from darkgraylib.tests.test_command_line import _make_test_argument_parser
